♻️ refactor [#55] 수정사항 적용
